feat: First pass attempt to support shared workers.
Changes to Canvas.tsx are straight forward. Changes to render.ts are simple and should work across a wide variety of situations, but may require some additional thought around ergonomics-- for the moment I am passing in a second, optional parameter to "render" that is the message port from the shared worker. A developer using the render function would presumably use "onconnect" in their shared worker and pass the message port into the render function.